LPS-54436 move to module
