+++ modules/portal/portal-executor/src/com/liferay/portal/executor/internal/PortalExecutorFactoryImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.executor;
+package com.liferay.portal.executor.internal;
 
 import com.liferay.portal.kernel.concurrent.RejectedExecutionHandler;
 import com.liferay.portal.kernel.concurrent.ThreadPoolExecutor;
